ci: Only build Flatpak for releases or on request
Previously, the Flatpak bundle was built for all commits on `main`, but
only if the repository is not a fork. I added the latter condition to
reduce what is almost certainly wasted computation.

Now that Flatpak bundles are being attached to releases, which we are
making every 2-3 weeks, I would now like to go further, and also disable
Flatpak builds on `main` by default. I don't think anybody is regularly
using them; they take several minutes and involve pulling gigabytes of
container image to run the build.

Instead, add an input to the workflow when manually dispatched to enable
the Flatpak job. This can be used to test changes to the Flatpak build
process.