Guard: Use body entered/exited signals for detection
This is for the "cone of light" detection, not the instant detection.
Although this one also has an option "instantly detected on sight" that
can be confusing.

Remove checking in every game update for the overlapping bodies of the
detection area. Use the body entered/exited signal instead for keeping
track of the player. This is assuming that there is a single player to
track, which was also the case in the previous implementation that was
taking get_overlapping_bodies().front().

Still do the raycast calculation in every game update if there is a
player that entered the area. This is to cover potential moving walls,
guard or player moving and something getting in the middle while the
player is still inside the detection area.

Also, do not update every time the exclamation sign progress bar in the
ALERTED state, because it doesn't change. Do it directly when entering
this state.