Ignore macOS AppleDouble files
In a learner's branch, I spotted a number of files whose names begin
with `._`, and `__MACOSX` directories. This is something to do with the
way that macOS stores extended file metadata.

https://en.wikipedia.org/wiki/AppleSingle_and_AppleDouble_formats

We do not want these files to be committed, just as we do not want the
related `.DS_Store` files, which were already listed in `.gitignore`
because they are present in the Tiny Swords archive.

Add `._*`, `__MACOSX` and `.AppleDouble` (which I have not seen yet, but
it doesn't hurt to preëmptively ignore it) to gitignore. Remove
`.aseprite` – I think the original intention was to not commit files
matching `*.aseprite`, but in fact we do commit these files and it is
useful to do so.