gitignore: Ignore zip and rar files
In a learner branch, I noticed .rar files being committed.

I don't think we want to have these in the repo. We have previously had
a .zip file in the repo, and *.zip is configured to be stored in git-lfs
by .gitattributes, but we removed that one file again in commit
12c6e14b14c5be197e78487ed551ebc4df401156.