tileset: Fix physics layer of void
In this tileset, physics layer 0 corresponds to layer 5 ("walls") in the
game, while layer 2 corresponds to project layer 10
("non_walkable_floor").

The void tiles incorrectly had their physics shape on layer 0, not 2.
This has not so far caused any practical problems but it would mean
that, for example, if void tiles were used in an ink combat level, ink
blobs would incorrectly bounce off the void, rather than floating over
it.

Update the layer for each of the 5 void assets.