Consolidate theme resources
For historical reasons, we have ended up with two theme resources. One,
in shared_components, is used by HUD input hints and the "Talk"/interact
label. The other, which is under the "dialogue" scene because it started
out as being used only for the dialogue balloons, is used by menus and
other panels in the game such as the credits screen.

Merge the two together. Take the smaller number of styles from
shared_components/theme.tres, and add them to
dialogue/components/theme.tres. Adjust all scenes that use the former to
use the latter. Then rename the latter over the former.

Fixes https://github.com/endlessm/threadbare/issues/381