tileset: Remove z_index -1 from water & foam
These were added in commit dd6ba63 ("Modified stone walls in tileset so
they work with lights and occlusion") but this is believed to be a
mistake. Arranging the TileMapLayers in the correct order should be
sufficient to stack water & foam correctly relative to other layers.
Setting a lower z_index on these tiles makes it harder to do other
tricks, like replacing the current "void" tiles with a parallax
starfield background behind all TileMapLayers.

This change requires corresponding changes in two scenes that use the
tileset:

- closing_transition: The "Void" TileMapLayer included some stray water
  tiles which, due to the order of TileMapLayer nodes, was now drawn on
  top of the grass. Remove these.

- stella_combat: The foam layers were incorrectly ordered above the
  sand. Move them to be above the water but below the sand.