closing_transition: Improve foam
This scene doesn't need to be perfectly drawn because only a small
region of it is visible in-game: it's a cut-scene where the camera only
shows a fixed area.

However I noticed that there was some missing foam within that fixed
area. Add it, and remove some stray foam from a previous level design
where (presumably) the island was smaller.