Use consistent color for InteractArea CollisionShape2Ds
Previously, these were a mix of yellowish colors, cyan for checkpoints,
and the default color.

Consistently use the same shade of yellow for all interact areas, as
well as the player's interact zone. This makes it a bit easier to
distinguish the interaction area from the collision area that prevents
walking through objects.