Add parallax effect for void starfield
We currently have a “void” tileset, which is used only in the closing
transition of the lorequest and as scenery in stella_stealth. Actually,
we have 5 tilesets with stars at different brightnesses, and a Void
terrain which combines them.

Conceptually, the stars are far away, so it would be nice to have a
parallax effect where the stars do not move even as the camera moves.

Implement this with a new variant of the void tiles plus a shader
applied to the TileMapLayer. The new tiles keep the existing stitched
edges, but then fill the rest of the space with pure red.

The shader uses red as its chroma key, replacing it with stars based on
a noise texture. The position sampled in the noise texture is not the
original UV for the fragment shader, but is instead based on the
viewport position (calculated in the vertex() stage and passed to the
fragment() stage).

A subtle flickering effect is added, using a sine function of the TIME
built-in shader variable (seconds since the engine started).

The new tiles have the same collision shape as the other 5 voids,
namely fully-opaque non_walkable_floor.

This implementation combines ideas from a variety of resources:

- The idea to use a chroma-key to identify the region to fill with stars
  is due to Manuel Quiñones. (My original idea was to make the void
  transparent, but this only works if the void is below all other layers.)

- The noise-texture starfield is inspired by a video by FencerDevLog
  <https://www.youtube.com/watch?v=EQv5kzqpYOk>, particularly the use of
  pow(r, 20.0) to compress most values in the noise texture towards
  zero.

- https://github.com/Ymanawat/Chroma-Key-Shader-Godot-4 was an
  interesting reference, though since this is a pixel-perfect texture
  there was no need to use the threshold + blending technique there.

- “Repeated texture overlay for tilemaps” by jess.codes
  <https://godotshaders.com/shader/repeated-texture-overlay-for-tilemaps/>
  showed me how to calculate a position in vertex() and pass it through
  to fragment(), though the calculation here is different because we
  want viewport position not world position. (As I write this I realise
  I could use the same mix() technique and avoid branching in the
  shader.)

https://github.com/endlessm/threadbare/issues/939