Closing transition: Delay fade-out a little
The walk-off animation is 2 seconds long, but currently once it starts
the dialogue ends without waiting for it to finish, and the Cinematic
immediately triggers the scene transition, which has a 1-second fade.

Instead, delay the end of the dialogue by 1 second, so that the scene
transition begins midway through the character's walk. This feels a
little less abrupt and is more symmetrical with the walk-on.