Main menu: Make root node a plain Control
Previously the root node was the HBoxContainer that lays out the logo to
the left of the menu items. Reparent it to a Control and adjust the
script accordingly.

This will make it easier to place a label with the version number in the
corner without interfering with the rest of the layout.