Support modulating trees
On entry to fragment(), COLOR = modulation * texture(TEXTURE, UV),
roughly. But the tree shader re-samples TEXTURE in fragment(),
overwriting COLOR. This means that the modulation is lost. In particular
this means you can't make a semi-transparent tree with the modulate
alpha channel.

In vertex(), COLOR is just the modulation. Save this in a varying, and
reapply it in fragment().