Player: Add optional run animation
If the SpriteFrames resource has an animation named "run", play a
matching "run" animation in the AnimationPlayer. It assumes that there
are 6 frames being played in sequence (like in the walk animation) and
it has an audio track for the running steps, playing the same audio FX
than the walk animation, but more often (at frames 0, 2 and 4).

If there is no "run" animation, the behavior stays the same as before:
play the walk animation at double the speed.

Note: As it happened in commit 48e35f24 the walk animation subresource
is moved again, so the diff is bigger than it should.

Fixes https://github.com/endlessm/threadbare/issues/953