Add TalkBehavior node
This node wraps few logic, but it is a reusable component that lets
developers easily trigger dialogues when the player interacts with an
area, with only setup and no scripting.

It inherits from Node because there is no 2D involved.

It is a tool script only because of the configuration warning.