Play from Fray's End when running from the Editor
Change the project main scene to be a new loader scene. The only thing
that this scene does is to switch to the Splash scene if the game is not
running in the editor, and to Fray's End if the game is running in the
editor.

Fray's End is the actual main scene of the game. For development, it is
better to skip the splash and intro and land directly in Fray's End.

Playing from the Splash scene is still possible by opening that scene in
the Editor and clicking the "Run Current Scene" button.