FixedSizeLabel: Throttle the _process virtual method
Before changing the position on screen, calculate the distance that it
will jump from the current position. Use that distance to change the
rate at which the next call to _process will happen. If the jump is big,
update it quickly. If the jump is small, take more time to update. Clamp
it for safety so the smallest jump takes at most
2 seconds to update, and the biggest jump takes at least 0.1
seconds.

This removes the dizzying effect of the label being repositioned while
the camera moves.

Helps https://github.com/endlessm/threadbare/issues/725