splash: Keep centred in viewport
Since commit 73b1ba691d540e95e2c749aaf5b30bdd8f9b6d12 ("Project: Expand
aspect ratio") the viewport may be wider or taller than 1920×1080. But
the logo was being drawn to a plain Node2D, using points that assumed a
1920×1080 viewport. Similarly the background was a Sprite2D.

Rebuild this scene to use a Control node at the root. Change the
background to a TextureRect, configured to Keep Aspect Covered.

Make the LogoStitcher node a Control rather than a Node2D, with a custom
minimum size of 1920×1080 (the resolution the points in the Path2D
assume) and reparent it to be centred in an AspectRatioContainer.

This keeps the logo centred in the viewport when the viewport is not at
16:9 aspect ratio.

(If we later change the base resolution of the game we will likely want
to revisit this.)

Fixes https://github.com/endlessm/threadbare/issues/995