Rock: Fix collision layer & mask
Previously the rock's collision layer and mask were both set to layer 1,
which is the player. Among other issues, this means that if a rock is in
the player detection area of a sign, the sign's label would always be
shown.

Fix the rock's layer to 5 ("wall") and mask to empty (it does not need
to detect collisions with anything).