Fray's End: Position Trees container at (0, 0)
Previously the Trees container that held all the trees in the scene was
positioned with a small offset (-1, 120). I think it was probably moved
accidentally.

Instead, position the container at (0, 0) and adjust all its children's
relative positions so that their absolute position remains the same. (I
did this by creating a new Node2D, dragging all the trees into it,
deleting the empty old container and renaming the new container.)
Lock the new container so it is harder to move it in future.