flatpak: Add sandbox hole to inhibit screensaver
Current Godot releases support using org.freedesktop.ScreenSaver to
inhibit the screensaver, but do not support using the Inhibit portal.

I have implemented Inhibit portal support upstream, but it is awaiting
review; and since it missed the 4.5 feature freeze it will not be part
of a Godot release until 4.6, which is probably close to a year away.

https://github.com/godotengine/godot/pull/108704