Add 12px margin to icon
This is to conform with Flathub's icon guidelines.

This was a little annoying to create because the grass tiles are 64×64.
I constructed this by taking the 93x93px of the top-left of the tilemap,
then adding the borders by hand from the tilemap, tweaking a little to
make it fit better. (The stitching at the edge of the tile is about 11px
wide/tall.)

I took the thimble-with-shadow from the previous icon by hand. I don't
think we have a reference copy of that.

https://docs.flathub.org/docs/for-app-authors/metainfo-guidelines/quality-guidelines#app-icon
https://github.com/endlessm/threadbare/issues/369