linux: Split release metadata to separate file
In theory, this allows the release information to be updated after the
release is published (e.g. to add translations, fix typos, etc.). Either
the software centre app on end-user machines, or the repo server when
generating the catalogue metadata, is meant to fetch the external data
and merge it into the file. (It is unclear to me whether this is meant
to happen client-side or on the repo server.)

https://www.freedesktop.org/software/appstream/docs/chap-Metadata.html#tag-releases