Add global node to debug aspect ratio issues
The project is set to expand the aspect ratio. This is to consider displays that
don't exactly match the aspect ratio of the base resolution (16:9). However we
expect that the game is played in a horizontal screen with more-or-less this
aspect ratio. Otherwise players could see unintended parts of the level and
level designers would have to consider corner cases, and artists wouldn't know
how far to dress the level.

Add minimum and maximum aspect ratios to Settings.

Add a global scene with a CanvasLayer overlay and a ColorRect that shows
semi-transparent red bars when the viewport is too tall or too wide. By playing
the game with this overlay, one can see the area that can be reached by the
camera and the forbidden zones.

The minimum and maximum aspect ratios in this commit are rather arbitrary.

Will Thompson had the idea of becoming this an autoload instead of playing the
whole game inside a subviewport, which removed all limitations.

Co-authored-by: Will Thompson <wjt@endlessaccess.org>