Void: Respect modulate in shader
On entry to fragment(), COLOR = modulation * texture(TEXTURE, UV),
roughly. But in the chroma-keyed region of the shader, COLOR is
overwritten (of course) which means that the modulation is lost.

In vertex(), COLOR is the modulation. Save this in a varying, and apply
it in fragment() after COLOR has been recalculated.

This allows modulating the void as a whole.

(This is the same approach as in commit d3273ffe, for trees.)