mailmap: Map @endlessos.org committers to @endlessaccess.org
We have changed the canonical domain name for our organisation.

It is not possible to edit the Author: and Committer: fields of old
commits, but it is possible to remap them in the output of git commands
to the new domain. Update all Endless Access team members who have
committed to date, including some of their other email addresses.