Music puzzle: Remove unused trees.gd script
This is a spiritual ancestor of the AreaFiller node. It randomises the
SpriteFrames of all its children to be one of two trees, hardcoded in
the script.

Nowadays, every tree in the music puzzle has the same (red) appearance;
and there are bushes that are children of the Trees node so setting them
to a tree sprite would be bad.

Remove this unused script. Random forests can be generated somewhat more
flexibly with AreaFiller these days.