Change void enemy dust particles
Change it to be an explossion that happens when tile cells are consumed. So
instead of having the particle node instantiated in the void spreading enemy and
constantly emitting, move it to its own scene and instantiate it each time it
has to explode.

Make the particle Explosiveness to 1 and and One Shot to true. So when
exploding, it will occlude the removed tile cells.

Change the return type of TileMapCover.consume_cells() to be a boolean that
tells if any cell was actually consumed.

Change the texture to be a sheet of 5 dust variants by setting a
CanvasItemMaterial with Particles Animation enabled. Then pick them randomly by
setting the ParticleProcessMaterial Display -> Animation -> Offset to be random
between 0 and 1. Animate the radial velocity and scale so each blob is slightly
different and they scale down while fading.