Set base resolution to 720p
Downscale the viewport size by two thirds, from 1080p to 720p. Because this is
the resolution of handheld consoles like Steam Deck or Nintendo Switch.

Also: change the window stretch mode from Viewport to Canvas Items. This is not
a pixel-perfect game. Otherwise by just downscaling the viewport size, the
background shows shakiness while the player (and thus the camera) moves.

Change the zoom of Camera2D nodes in scenes depending on their content. For the
Lore quests:

- Fray's End, Music Puzzle, Stealth Level, Void Runner: reset zoom to 1 because
  these are open areas. Note that the area framed around the player in these
  scenes is bigger now. Same for other open areas that aren't part of the lore
  quests like the test scenes.

- Sokobans: reset zoom to 1 in all of them. This adds more or less dark area
  around depending on the previous zoom.

- Ink Combat: set zoom to 2/3 of the previous value to keep the zoom as before
  in levels 1 and 2. Reset zoom to 1 in level 3 because the convertion gives
  almost one (1.066). Note that there is a row of tiles missing now on the right
  side!

Template, Stella and "El juguete perdido" StoryQuests:

- Stealth: reset zoom to 1.

- Intro, Outro, Combat, Sequence puzzle: set zoom to 2/3 of the previous value
  to keep the zoom as before.

Resolves https://github.com/endlessm/threadbare/issues/638