Walk behaviors: Factor out input speed parameters
Previously, each behavior defined a `walk_speed` property, with the input
behavior also defining `run_speed` and parameters controlling acceleration and
deceleration. It seems likely that we will in due course want the
non-input-controlled walk behaviors to have these other parameters, to give them
more natural movement.

Define a new CharacterSpeeds resource that holds the 4 parameters from
InputWalkBehavior. Adjust all walk behaviors to have a `speeds` property.

The property defaults to `null`; it is set to an instance with the default
values on ready in that case, but only when running in-game. This is because we
do not want to persist the default values in each scene where this node is used,
only if they have been changed; and there is no way in GDScript to define a
constant instance of CharacterSpeeds. (If we defined a
default_character_speeds.tres and set this as the default, then it would be very
easy to accidentally edit the global defaults.)

See discussion at
https://github.com/endlessm/threadbare/issues/1151#issuecomment-3253099009.