Walk behaviors: Move stuck_speed to CharacterSpeeds
This is not used by InputWalkBehavior, but it is used by all the non-interactive
walking behaviors, and you could imagine it being used by the InputWalkBehavior
in future, perhaps to show a "pushing" animation.

It is a little awkward to phrase the documentation because the result of the
character being stuck is different between the different behaviors; I tried to
move the key differences to elsewhere in each behavior's documentation.

I added a helper method is_stuck() primarily because the code that was
previously inlined ends up being wrapped across 4 lines by gdformat, but I think
it's clearer anyway.