dialogue: Add problematic glyphs to balloon placeholder text
Our old font, m6x11plus, does not have the ’ glyph. While it is hard to type
this directly into Godot Dialogue Manager (if you don't know the keyboard
shortcuts to type fancy quotes), people often write dialogue in a word processor
(which automatically converts ' to ’) then paste it into Godot.

Replace the straight apostrophe ' with a curly apostrophe ’ in the placeholder
text in the balloon scene; and add some other glyphs that we expect to be used
in dialogue which may not be in all fonts. This makes it easier to see the
result when experimenting with alternative fonts.