theme: Switch to Jersey font
The sizes in the Jersey variants' names refer to the height of capital letters,
in pixels. The sizes used by Godot are the full height of the font, which
includes padding. I found the correct size values by trial and error.

This preserves the approximate size of text in the game. We will want to change
this in future, given the switch to 720p base resolution. Jersey has more size
variations than m6x11plus, so this will help us to make this switch.

Fixes https://github.com/endlessm/threadbare/issues/201