assets: Add Jersey font
This is a pixel font designed for knitting. As well as looking good and fitting
the game thematically, it has better glyph coverage than our current font, and
has a greater variety of sizes available.

OFL-1.1 is a free license but requires the license to be displayed in-game. I
will add this to the credits page in a follow-up commit.

The smallest size, Jersey10, is configured with antialiasing and
subpixel rendering disabled. This matches the import settings for
m6x11plus. The larger sizes, to my eye, look better with these enabled,
so I kept the defaults; we can adjust this later if better eyes
disagree!

https://github.com/endlessm/threadbare/issues/201