theme: Reduce font sizes
When we switched the game's base resolution from 1920×1080 to 1280×720, we did
not adjust the font size to compensate.

The new resolution is 2/3 of the old resolution in both axes. The previous font
only had one size, and we were using it at 2× that size. Halving it to its
natural size looked too small against the scroll assets.

Having switched to a font that has a variety of sizes, we can now reduce its
size by approximately 1/3:

- The default font was Jersey-20 at 1×. Change it to Jersey-15. Since Jersey-20
  is slightly smaller than the old font was, this works out at almost exactly
  2/3 the size of m6x11plus.

- BigFlatButton (used for the buttons on the main menu) was Jersey-20 at 2×.
  Change it to Jersey-25 at 1×.

Helps https://github.com/endlessm/threadbare/issues/1136