dialogue: Reduce balloon size & spacing
Since changing the game's base resolution, dialogue balloons have taken
up way too much of the viewport. Reducing the font size and box
margins/padding have helped a bit but we can go further:

- Reduce the minimum size of the balloon by 1/3 in the y-axis and 1/5 in the
  x-axis. To my eye, reducing the minimum width to 320 looked too narrow.

- Reduce the padding around the Next button by eliminating the HBoxContainer and
  MarginContainer it was packed in. Adjust its horizontal container sizing to
  shrink end to preserve its right-aligned position; and the script to show and
  hide the button rather than the now-removed container.

The assets from Tiny Swords that we are using for the balloons (and other UI
elements) are the limiting factor in making this much smaller, since they are
based around a 64×64 grid. In future I think we'll want to replace them.