theme: Reduce transparent margin around PanelContainer
The scroll/parchment texture we are using for PanelContainer's background has
around 30px of transparent border. For PanelContainer we would prefer to use
this space as margin between the edge of the visible parchment and its contents.

Expand the margin by 16px in each axis. Reduce the content margin at the top by
16px to compensate - the shape of the scroll means there is additional margin
there.

Co-authored-by: Will Thompson <wjt@endlessaccess.org>