dialogue: Reduce minimum balloon size
Due to the previous change to effectively add 16px to the content
margins for PanelContainer, we can reduce the minimum width of the
balloon by 2×16px to compensate.

In addition, hide the ResponsesMenu container, and update the offsets
saved to the PanelContainer node in the scene. Previously the saved size
included both the "Next" button and 1 row of responses; now, the saved
size has space for the speaker's name, 3 lines of dialogue, and a next
button. This means that in most cases, the balloon does not need to grow
when the "Next" button appears or between lines of dialogue.