Grappling hook: Add HookControl and HookableArea
These new nodes are part of the grappling hook mechanic.

HookableArea extends Area2D to connect the grappling hook.

HookControl handles the input to control the aiming and throwing of the
grappling hook.

Add a new global group hook_listener to the project. Because the
grappling hook can go across multiple connections, it is easier to use
call_group() for communication rather than emitting a signal, capture it
and re-emitting it multiple times.

Add a new hookable collision layer. The HookableArea will warn in the
editor if it doesn't belong to it. Also provide a helper scene with just
the HookableArea added to the corresponding collision layer. Like the
one we have for InteractArea.

Also add an arrow sprite asset for the HookControl.