Grappling hook: Add it to player
Add a new HOOKING mode to player. Use it to enable a new PlayerHook
node that contains the main logic of the grappling hook tool. This tool
makes use of the two classes added in the previous commit: It has a
primary HookControl, and it keeps an array of the HookableAreas that
were connected. It also draws the Line2D of the string and moves the
player and the hooked game element when pulling.

Export a new aiming_speed in player. Used to slow down the player
movement while aiming with keyboard.

In the player scene, move the HitBox node outside PlayerFighting, so it
can be used in other modes. For instance, to collect items.

Use the attack_01 frame-by-frame animation when throwing the string.