Add HookableButtonItem
A CharacterBody2D that can be hooked with the grappling hook. It has a
ButtonItem instantiated with editable children to change the sprite to a
different variation (purple).

It also has a HookableArea with a Marker2D to indicate where to connect
the string. And set the weight to zero. With this, the button can be
reached and collected if it's over water or void.

Set the debug color of the HookableArea shape to purple. And make this
consistent in forthcoming commits, so all purple areas are hookable.