Add HookablePin and HookableNeedle
The pin can be hooked to cross a river or void. Since it is not a
CharacterBody2D the player moves to it when it's hooked.

The needle is a connection, its HookableArea has a HookControl so the
player aims again when connecting it.

Add sprites for both game elements.