Factor NavigationFollowWalkBehavior out of void_spreading_enemy.gd
Previously, void_spreading_enemy.gd contained movement logic somewhat
entwined with the other logic for this entity. But the movement logic
was relatively generic – follow a target using a NavigationAgent2D – and
you could imagine a different flavour of this entity that moves in a
different way, perhaps following a path.

Create NavigationFollowWalkBehavior based on the logic in
void_spreading_enemy.gd, with similar functionality to
FollowWalkBehavior but using a NavigationAgent2D rather than
straight-line pathfinding.

As with the previous implementation, if the distance to the target is
above a configurable threshold, the character runs; otherwise, they
walk. Like InputWalkBehavior, a running_changed signal is emitted, which
can be connected to CharacterSpriteBehavior to control animations,
though this is not used in the void runner scene.

Signals are emitted when navigation ends (either indicating that the
target has been reached, or that it cannot be reached), though this is
also not used in the void runner scene.

Fixes https://github.com/endlessm/threadbare/issues/1175